% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicurve.R
\name{epicurve}
\alias{epicurve}
\title{Plot the epidemic curve}
\usage{
epicurve(
  incidence,
  dates = NULL,
  datelab = "7d",
  col = "deepskyblue4",
  barwidth = 1,
  title = "Epidemic curve",
  xtickangle = 0
)
}
\arguments{
\item{incidence}{A vector for the time series of incidence.}

\item{dates}{A vector of dates in format "YY-MM-DD".}

\item{datelab}{The spacing for ticks on the x-axis. Default "7d".}

\item{col}{The color of the epidemic curve.}

\item{barwidth}{The width of the bars. Default is 1.}

\item{title}{Title of the plot.}

\item{xtickangle}{The angle of the x-ticks. Default is 0 (horizontal).}
}
\value{
A plot of the epidemic curve based on the provided time series of
counts.
}
\description{
This routine gives a graphical representation of the epidemic curve based
on case incidence data.
}
\examples{
si <- c(0.05, 0.05, 0.1, 0.1, 0.1, 0.1, 0.1, 0.05, 0.05, 0.1, 0.1, 0.1)
epidemic <- episim(serial_interval = si, Rpattern = 4)
epicurve(epidemic$y)
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
