% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episim.R
\name{episim}
\alias{episim}
\title{Simulation of epidemic data based on Poisson counts}
\usage{
episim(serial_interval, endepi = 50, Rpattern = 1, verbose = FALSE, plotsim = FALSE)
}
\arguments{
\item{serial_interval}{A vector of values for the serial interval.}

\item{endepi}{The total number of days of the epidemic.}

\item{Rpattern}{Different scenarios for the true underlying curve of
Rt. Four scenarios are possible with 1,2,3,4.}

\item{verbose}{Should metadata on simulated edpidemic be printed?}

\item{plotsim}{Create a plot of the incidence time series, the true
reproduction number curve and the serial interval.}
}
\description{
Based on a serial interval and a functional input for the reproduction number
over T days, the routine generates a set of incidence counts following a
Poisson model. The link between the reproduction number and the generated
incidence data is governed by the renewal equation. The baseline mean number
of cases at day 1 is fixed at 15. The mean number of cases for the remaining
days of the epidemic are generated following equation (2) of Azmon et al.
(2013).
}
\examples{
si <- c(0.05, 0.05, 0.1, 0.1, 0.1, 0.1, 0.1, 0.05, 0.05, 0.1, 0.1, 0.1)
epidemic <- episim(serial_interval = si, Rpattern = 1)

}
\references{
Azmon, A., Faes, C., Hens, N. (2014). On the estimation of the
reproduction number based on misreported epidemic data. \emph{Statistics in
medicine}, \strong{33}(7):1176-1192. \url{https://doi.org/10.1002/sim.6015}
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
