% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eruptions.R
\docType{data}
\name{eruptions}
\alias{eruptions}
\title{Eruption times in Yellowstone National Park}
\format{
\describe{
 \item{\code{eruptions}}{A vector of eruption times in minutes.}
}
}
\usage{
data(eruptions)
}
\description{
A vector of eruption times in minutes recorded from the Old
 Faithful geyser in Yellowstone National Park, USA.
}
\references{
Härdle, W. (1991). Smoothing Techniques with Implementation in
S. New York: Springer.
}
\keyword{datasets}
