% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimRmcmc.R
\name{estimRmcmc}
\alias{estimRmcmc}
\title{Estimation of the reproduction number with Laplacian-P-splines via MCMC}
\usage{
estimRmcmc(incidence, si, K = 30, niter = 5000, burnin = 2000, dates = NULL,
 CoriR = FALSE, WTR = FALSE)
}
\arguments{
\item{incidence}{A vector containing the time series of incidence counts.}

\item{si}{The discrete serial interval distribution.}

\item{K}{Number of B-splines in the basis.}

\item{niter}{Number of iterations for the MCMC sampler.}

\item{burnin}{The burnin size.}

\item{dates}{A vector of date values (optional).}

\item{CoriR}{Should the Rt estimate be also computed with the method of Cori?}

\item{WTR}{Should the Rt estimate be also computed with the Wallinga-Teunis
approach?}
}
\value{
An object of class \code{epilps} containing the pointwise and set
 estimates of the time-varying reproduction number and the epidemic curve
 respectively.
}
\description{
This routine estimates the instantaneous reproduction number Rt (the mean number
of secondary cases generated by an infectious individual at time t, White et
al. 2020) using Bayesian P-splines and Laplace approximations. The inference
approach is fully stochastic with a Metropolis-adjusted Langevin algorithm. The
\code{estimRmcmc()} routine estimates Rt based on a time series of incidence
counts and a given serial interval distribution. The negative
Binomial distribution is used to model incidence count data and P-splines
(Eilers and Marx, 1996) are used to smooth the epidemic curve. The link
between the epidemic curve and the reproduction number is established via the
renewal equation. If \code{incidence} contains NA values at certain time
points, these are replaced by the average of the left- and right neighbor
counts. If the right neighbor is NA, the left neighbor is used as a
replacement value.
}
\examples{
si <- c(0.05, 0.05, 0.1, 0.1, 0.1, 0.1, 0.1, 0.05, 0.05, 0.1, 0.1, 0.1)
epidemic <- episim(serial_interval = si, Rpattern = 2, endepi = 30)
epifit <- estimRmcmc(incidence = epidemic$y, si = si)
plot(epifit)

}
\references{
Gressani, O., Wallinga, J., Althaus, C. L., Hens, N. and Faes, C.
 (2022). EpiLPS: A fast and flexible Bayesian tool for estimation of the
 time-varying reproduction number. \emph{Plos Computational Bioloby},
 \strong{18(10): e1010618}.

White, L.F., Moser, C.B., Thompson, R.N., Pagano, M. (2021).
 Statistical estimation of the reproductive number from case notification
 data. \emph{American Journal of Epidemiology}, \strong{190}(4):611-620.

Eilers, P.H.C. and Marx, B.D. (1996). Flexible smoothing
 with B-splines and penalties. \emph{Statistical Science},
 \strong{11}(2):89-121.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
