% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimSI.R
\name{estimSI}
\alias{estimSI}
\title{Nonparametric serial interval estimation}
\usage{
estimSI(x, B = 5000, verbose = TRUE)
}
\arguments{
\item{x}{A data frame with two columns based on transmission pair data. The
number of rows is the number of transmission pairs (sample size). The first
column should be the left boundary of the SI window and the second column
should be the right boundary of the SI window. NA values are not permitted.}

\item{B}{The number of bootstrap samples.}

\item{verbose}{Print basic summary features of the SI distribution?}
}
\value{
A list containing basic summary features with point estimates and
interval estimates.
}
\description{
The serial interval (SI) is defined as the time between illness onset in
a primary case (infector) and illness onset in a secondary case (infectee).
This routine estimates different features of the SI distribution based on
transmission pair data. The input \code{x} should be a data frame with two
columns containing the left and right boundary of the serial interval window
for each transmission pair. A resampling technique (bootstrap) is used to
draw samples from a nonparametric estimate of the cumulative distribution
function of the serial interval.
}
\examples{
data(influenza2009serial)
SIfit <- estimSI(x = influenza2009serial)

}
\references{
Gressani, O. and Hens, N. (2024). Nonparametric serial interval
estimation. \emph{MedRxiv preprint}.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
