% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Idist.R
\name{Idist}
\alias{Idist}
\title{Density function and discrete distribution for a disease interval}
\usage{
Idist(mean, sd, dist = c("gamma", "weibull", "lognorm"), probs = NULL)
}
\arguments{
\item{mean}{The mean of the disease interval (must ber larger than 1).}

\item{sd}{A positive and finite real number corresponding to the standard
deviation of the disease interval.}

\item{dist}{A choice among a Gamma, Weibul or Log-normal distribution for
the disease interval.}

\item{probs}{A vector of probabilities for the interval distribution.}
}
\value{
A list of class \code{Idist} containing a vector of probabilities
corresonding to the discrete distribution of the disease interval,
the name of the chosen parametric distribution and its parameters.
}
\description{
This function computes the probability density function and probability mass
function for a disease interval (e.g. the generation interval defined as the
time delay between infections in an infector-infectee pair). It takes as
input the mean and the standard deviation of the disease interval (usually
expressed in days) and gives as an output the interval distribution based on
a chosen parametric family (Gamma, Weibull or Log-normal).
}
\details{
The discretization is based on the formula in Held et al. (2019). The
probability mass function \eqn{g(t)} for the discrete time variable
\eqn{t \in \{1,\dots,D_{\text{max}}\}} is given by:

\eqn{g(t) = \frac{F(t+0.5)-F(t-0.5)}{F(D_{\text{max}}+0.5)-F(0.5)}},

where \eqn{D_{\text{max}}} is the greatest integer less than or equal
to \eqn{F^{-1}(0.9999)}, and \eqn{F} is the cumulative distribution function
of the chosen parametric distribution for the disease interval.
}
\examples{
Idist(mean = 2.6, sd = 1.5)
}
\references{
Held, L., Hens, N., D O'Neill, P., and Wallinga, J. (2019).
Handbook of infectious disease data analysis. \emph{CRC Press}.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
