% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incubsim.R
\name{incubsim}
\alias{incubsim}
\title{Simulation of incubation times}
\usage{
incubsim(incubdist = c("LogNormal","Weibull","MixWeibull", "Gamma"),
coarseness = 1, n = 100, tmax  = 20, tgridlen = 500, plotsim = FALSE)
}
\arguments{
\item{incubdist}{The distribution of the incubation period.}

\item{coarseness}{The average coarseness of the data.}

\item{n}{The number of observations.}

\item{tmax}{The upper bound on which to evaluate the \code{incubdist} density.}

\item{tgridlen}{The number of grid points on which to evaluate the density.}

\item{plotsim}{Create a plot of the simulated data.}
}
\value{
An object of class \code{incubsim} consisting of....
}
\description{
This routine simulates incubation times for a given generation times density
and incubation times density.
}
\examples{
incubsim()

}
\references{
Ferretti, L., Wymant, C., Kendall, et al. (2020). Quantifying
SARS-CoV-2 transmission suggests epidemic control with digital contact
tracing. \emph{Science}, \strong{368}(6491), eabb6936.

Backer, J. A., Klinkenberg, D., & Wallinga, J. (2020). Incubation
period of 2019 novel coronavirus (2019-nCoV) infections among
travellers from Wuhan, China, 20–28 January 2020. \emph{Eurosurveillance},
\strong{25}(5), 2000062.

Donnelly, C. A., Ghani, A. C., Leung, G. M., et al. (2003).
Epidemiological determinants of spread of causal agent of severe acute
respiratory syndrome in Hong Kong. \emph{The Lancet}, \strong{361}(9371),
1761-1766.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
