% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influenza2009serial.R
\docType{data}
\name{influenza2009serial}
\alias{influenza2009serial}
\title{Serial interval data for the 2009 outbreak of pandemic Influenza A (H1N1)}
\format{
A data frame with two columns:
\describe{
 \item{\code{SL}}{The left boundary of the serial interval window.}
 \item{\code{SR}}{The right boundary of the serial interval window.}
}
}
\source{
\url{https://www.nejm.org/doi/full/10.1056/NEJMoa0906089}
}
\usage{
data(influenza2009serial)
}
\description{
H1N1 Influenza A illness onset data for 16 transmission pairs
at a New York City School. The dataset is obtained from the supplementary
appendix of Lessler et al. (2009).
}
\references{
Lessler, J., Reich, N. G., Cummings, D. A., & New York City
Department of Health and Mental Hygiene Swine Influenza
Investigation Team. (2009). Outbreak of 2009 pandemic influenza A
(H1N1) at a New York City school. New England Journal of Medicine,
\strong{361}(27), 2628-2636.
}
\keyword{datasets}
