% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_nowcast_plot.R
\name{plot.nowcasted}
\alias{plot.nowcasted}
\title{Plot of nowcasted cases and delay distribution}
\usage{
\method{plot}{nowcasted}(x, type = c("cases", "delay"), ...)
}
\arguments{
\item{x}{An object of class \code{nowcasted}.}

\item{type}{Either "cases" (default) to show the nowcasted cases or
"delay" to show the contour plot of the estimated delay distribution.}

\item{...}{Further arguments to be passed to plot.}
}
\value{
A plot of the reported and nowcasted cases.
}
\description{
This routine plots the reported cases, the nowcasted cases and
associated \eqn{95\%} credible interval for the nowcast. In addition, it
can be used to show the contour plot of the estimated delay distribution.
}
\author{
Bryan Sumalinab (writing) and Oswaldo Gressani (editing).
}
