% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sars2003.R
\docType{data}
\name{sars2003}
\alias{sars2003}
\title{Daily incidence of the 2003 SARS epidemic in Hong Kong}
\format{
A list with two components:
\describe{
 \item{\code{incidence}}{A vector with 107 observations.}
 \item{\code{si}}{A vector of probabilities corresponding to the serial
 interval distribution.}
}
}
\source{
\url{https://cran.r-project.org/package=EpiEstim}
}
\usage{
data(sars2003)
}
\description{
A list with the daily incidence of onset of symptoms for the 2003
SARS outbreak in Hong Kong and a discretized serial interval distribution.
}
\references{
Cori A. et al. (2009). Temporal variability and social
heterogeneity in disease transmission: the case of SARS in Hong Kong.
PLoS Comput Biol 5(8) : e1000471.
}
\keyword{datasets}
