% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_Rt_summary.R
\name{summary.Rt}
\alias{summary.Rt}
\title{Summarize the estimated reproduction number}
\usage{
\method{summary}{Rt}(object, ...)
}
\arguments{
\item{object}{An object of class \code{Rt}.}

\item{...}{Further arguments to be passed.}
}
\value{
A summary output.
}
\description{
This routine can be used to summarize estimation results for
related to the reproduction number.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
